Add files via upload
Question:
A Basket is given to you in the shape of matrix .if the size matrix N*N then the range
of no. of eggs you can put in each slot of matrix is 1 to N**2. task is to arrange the eggs in the basket such that the
sum of each row ,column and the diagonal of the matrix remain same...